+++ src/types.ts
@@ -1,0 +1,7 @@
+import { Component, FC } from 'react';
+
+export type PropsOf<T> = T extends FC<infer P>
+	? P
+	: T extends Component<infer P>
+	? P
+	: never;