+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - ... <!-- Add new lines here. Version number will be decided later -->
 - feat: Make `Table` skip empty thead and tfoot sections
 - feat(ts): Export types `TableCellData` & `TableCellMeta`
+- feat(ts): Add `PropsOf<ComponentType>` utility type
 - fix: `Table`'s cell-level settings should override "cols" meta data
 
 ## 0.1.4 – 0.1.9